requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
django-elasticsearch-dsl==8.0
django-filter==25.1
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
phonenumbers==9.0.2
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
